Merger PlayerTests
I merged my tests with passer's
