+++ backend/src/test/java/com/catan/model/PlayerTests.java
@@ -1,5 +1,6 @@
-package com.catan;
+package com.catan.model;
 
+import org.junit.jupiter.api.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
@@ -8,11 +9,6 @@ import static org.junit.jupiter.api.Assertions.*;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
-
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.Test;
 
 import com.catan.exceptions.UserNotFoundException;
 import com.catan.model.Color;
@@ -68,6 +64,32 @@ public class PlayerTests {
         repositoryPlayer.deleteAll();
         repositoryPlayResCard.deleteAll();
         repositoryGame.deleteAll();
+    }
+
+    @Test
+    public void testPlayerInitialization() {
+        // Create a sample user
+        User user = new User();
+        user.setId(1);
+
+        // Create a player
+        Player player = new Player();
+        player.setId(10);
+        player.setUser(user);
+        player.setColor(Color.RED);
+        player.setVictoryPoints(5);
+        player.setNumberOfRoads(4);
+        player.setNumberOfVillages(3);
+        player.setNumberOfCities(2);
+
+        // Perform assertions
+        Assertions.assertEquals(10, player.getId());
+        Assertions.assertEquals(user, player.getUser());
+        Assertions.assertEquals(Color.RED, player.getColor());
+        Assertions.assertEquals(5, player.getVictoryPoints());
+        Assertions.assertEquals(4, player.getNumberOfRoads());
+        Assertions.assertEquals(3, player.getNumberOfVillages());
+        Assertions.assertEquals(2, player.getNumberOfCities());
     }
 
     @AfterEach
